remove legacy datastore mono helpers
